[core] Hostname logic refactoring
This commit:
- Moves hostname logic in it’s own util module
- Moves cloud metadata logic in it’s own util module
- Add a new way to determine the hostname using the kubelet
- Some code cleaning/consolidating
