+++ checks.d/tokumx.py
@@ -18,7 +18,6 @@ from pymongo import (
 
 # project
 from checks import AgentCheck
-from util import get_hostname
 
 DEFAULT_TIMEOUT = 10
 
@@ -244,7 +243,6 @@ class TokuMX(AgentCheck):
                 return 'Rollback'
 
         status = get_state_description(state)
-        hostname = get_hostname(agentConfig)
         msg_title = "%s is %s" % (server, status)
         msg = "TokuMX %s just reported as %s" % (server, status)
 
@@ -253,7 +251,7 @@ class TokuMX(AgentCheck):
             'event_type': 'tokumx',
             'msg_title': msg_title,
             'msg_text': msg,
-            'host': hostname
+            'host': self.hostname
         })
 
     def _get_ssl_params(self, instance):
