+++ ddagent.py
@@ -28,6 +28,7 @@ import zlib
 os.umask(022)
 
 # 3p
+import simplejson as json
 try:
     import pycurl
 except ImportError:
@@ -51,12 +52,11 @@ from config import (
 import modules
 from transaction import Transaction, TransactionManager
 from util import (
-    get_hostname,
-    get_tornado_ioloop,
     get_uuid,
-    json,
     Watchdog,
 )
+
+from utils.hostname import get_hostname
 from utils.logger import RedactedLogRecord
 
 
@@ -515,7 +515,7 @@ class Application(tornado.web.Application):
         log.info("Listening on port %d" % self._port)
 
         # Register callbacks
-        self.mloop = get_tornado_ioloop()
+        self.mloop = tornado.ioloop.IOLoop.current()
 
         logging.getLogger().setLevel(get_logging_config()['log_level'] or logging.INFO)
